Add zk-cred-longfellow-fields crate
Moves the derives field element implementations into a distinct crate,
and creates a cargo workspace so that the main crate can depend on it.
This significantly speeds up incremental compilation (e.g. in
rust-analyzer) of the main crate.

The downside is that I had to introduce a `struct Byte(u8)` because
otherwise the compiler complains that the implementation of `Codec` for
`u8` conflicts with the blanket implementation of `Codec` for trait
`FieldElement`. Even worse, this forces some copies when serializing and
deserializing, because we have to wrap/unwrap `u8`s into `Byte`.
