+++ Cargo.toml
@@ -16,4 +16,10 @@ rand = "0.8.5"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = { version = "1.0" }
 sha2 = "0.10.9"
+zk-cred-longfellow-fields = { version = "0.1.0", path = "fields"}
 zstd = "0.13.3"
+
+[workspace]
+# Deriving ff::PrimeField takes a long time, so fields is a distinct crate to speed up incremental
+# builds.
+members = [".", "fields"]
