+++ fields/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "zk-cred-longfellow-fields"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+anyhow = "1"
+byteorder = "1.5.0"
+ff = { version = "0.13", features = ["derive"] }
